requirements.txt
django
djangorestframework
markdown
django-filter
django-oauth-toolkit
django-cors-headers
numpy
django-suit
django-ckeditor
Pillow
